fix bedita cleanup on frontends

+++ bedita-app/vendors/shells/bedita_base.php
@@ -128,18 +128,20 @@ class BeditaBaseShell extends Shell {
 		}
     }
 
-    protected function __clean($path) {
+    protected function __clean($path, $removeDirs=true) {
         
         $folder = new Folder($path);
         $list = $folder->ls();
 
-        foreach ($list[0] as $d) {
-        	if($d[0] != '.') { // don't delete hidden dirs (.svn,...)
-	        	if(!$folder->delete($folder->path.DS.$d)) {
-	                throw new Exception("Error deleting dir $d");
-	            }
-        	}
-        }
+		if($removeDirs) {
+	        foreach ($list[0] as $d) {
+	        	if($d[0] != '.') { // don't delete hidden dirs (.svn,...)
+		        	if(!$folder->delete($folder->path.DS.$d)) {
+		                throw new Exception("Error deleting dir $d");
+		            }
+	        	}
+	        }
+		}
         foreach ($list[1] as $f) {
         	$file = new File($folder->path.DS.$f);
         	if(!$file->delete()) {
